Refactor toolbar sorting into components folder
