+++ src/locale/misc/en.yml
@@ -25,3 +25,9 @@ nativePersonFields:
   zip_code: Zip code
 snackbar:
   error: Oh dear, something went wrong
+dataTable:
+  sorting:
+    addButton: Add sorting column
+    button: Sort
+    hint: 'Hint: hold down {shiftKeyIcon} while clicking multiple columns'
+    title: Sort
